+++ src/lib.rs
@@ -244,7 +244,7 @@
 //! - `"const_marker"`(enabled by default): enables the [`const_marker`] module,
 //! and all items that depend on it.
 //! 
-//! - `"nightly_const_marker"`(requires the nightly compiler):
+//! - `"adt_const_marker"`(requires the nightly compiler):
 //! enables the `"const_marker"` feature,
 //! and marker types in the [`const_marker`] module that have
 //! non-primitive `const` parameters.
@@ -279,8 +279,8 @@
 //! [`const_marker`]: crate::const_marker
 #![no_std]
 #![cfg_attr(feature = "nightly_mut_refs", feature(const_mut_refs))]
-#![cfg_attr(feature = "nightly_const_marker", feature(adt_const_params))]
-#![cfg_attr(feature = "nightly_const_marker", allow(incomplete_features))]
+#![cfg_attr(feature = "adt_const_marker", feature(adt_const_params))]
+#![cfg_attr(feature = "adt_const_marker", allow(incomplete_features))]
 #![cfg_attr(feature = "docsrs", feature(doc_cfg))]
 #![allow(clippy::type_complexity)]
 #![deny(missing_docs)]
@@ -307,7 +307,7 @@ pub mod type_fn;
 #[cfg_attr(feature = "docsrs", doc(cfg(feature = "const_marker")))]
 pub mod const_marker;
 
-#[cfg(feature = "nightly_const_marker")]
+#[cfg(feature = "adt_const_marker")]
 mod all_init_bytes;
 
 mod utils;